+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Changelog
 **Improvement**
 
 * Better display of **xfailed** and **xpassed** multi-line reason messages.
+* Failed test teardowns are included in the report.
 
 
 1.3.7